requirements.txt
crewai
python-dotenv
langchain_community
